src/third_party/blink/web_tests/virtual/out-of-blink-frame-ancestors-disabled/external/wpt/portals/csp/frame-ancestors.sub-expected.txt

This is a testharness.js-based test.
FAIL Violation report status OK. assert_equals: No such report. expected "" but got "false"
Harness: the test ran to completion.
